correccion pelispedia y animeyt con nuevo gktools
